+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Keep a local seismic catalog.
 
 Copyright (c) 2022-2024 Claudio Satriano <satriano@ipgp.fr>
 
-## unreleased
+## v0.8 - 2024-10-28
 
 - New option for `seiscat print` and `seiscat plot`: `--where` to filter events
   based on one or more conditions